add the option to register a single command instead of a handler
